Change raise_errors, ini file handling and fix docs
This changes raise_errors to default to True. This handles the
SECRET_KEY use case better and reduces the likelihood of error. This
also adds an example to the module documentation about it.

This changes ini file handling such that if you provide ANTENNA_INI
in the environment, then it'll try to open that fail and if it fails,
it'll raise a ConfigurationError. This also leads to fewer errors.