Add datetimeutil module and tests from socorrolib
Antenna will be self-contained and won't depend on anything from socorro
or socorrolib. However, there are a few things we need to use from
socorrolib to make everything work.

As such, my plan is to copy code from socorrolib that's very very
unlikely to change.

This copies over the datetimeutil code and its tests. It also updates
both so they don't make our linter sad and also to take advantage of
pytest magic.