Remove basic auth code
We're going to do the basic auth check in nginx rather than do it in
the web app itself. That's a much better idea.

This removes all the basic auth code and tests.

Further, it updates WebTest to the latest version and it fixes a
couple more Python 3 issues so that the tests now pass.