Flesh out breakpad handling; tweak test infrastructure
This fleshes out some of the breakpad crash report handling, but it doesn't
work, yet.

This also creates some test scaffolding to make it easier to test GET
and POST requests with.

Falcon 1.0 has some niceties for testing, but I found it inflexible
when used with pytest. They've redone a lot of that since the 1.0
release, so when they release a new version and we upgrade, we can
redo the test client stuff.