Implement FSCrashStore
This implements a rudimentary file system crash store that mimics the
"directory structure" of what we're going to do on S3, but does it on
a local file system.

There are a bunch of FIXMEs that should get thought through at some
point, but since I'm only using this crash store for
development/testing purposes, I'm going to leave those be for now.