Implement s3 crash storage
This implements the s3 crashstorage system and also an s3mock which
makes it easier to test the s3 functionality.

Additionally, I made some changes to the docker environment adding a
base container which doesn't start up the fakes3 container. That's a
better place to run tests, lint, and other things that don't require
fakes3.