Implment multipart_encode
We're doing a lot of stuff with HTTP POST multipart/form-data
payloads, so it behooves us to have a way to generate those in a
convenient way such that we can write tests that are easy to verify
and maintain.

This will help us writing tests and other utilities we'll need to
develop and maintain antenna.