Implement retry decorator
This implements a retry decorator that lets you retry for specified
exceptions as well as function return values.

It allows you to specify the sleep function so we can use gevent's
sleep.

It allows you to specify max attempts.

It has retry amounts that increase up to 2 minutes and keeps that
number--that matches what we want to be doing in production.

It will log all the exceptions if passed a logger so we can more
easily discover when it's gone awry.