Fix docs creation
This creates a bunch of docker stuff so that we can have a separate
docker image for doing development-specific things without having some
of the tools we need to install available in the deploy image.

The new local/antenna_dev image has make which we need to build the
docs.