Implement metrics infrastructure
This implements infrastructure for metrics. It's global to the entire
app to make it easier to add metrics easily without having to add
additional tangling between things. The theory is that metrics is like
logging where it's available to the entire app because we always want
it.

This adds a LoggingMetrics and a DogStatsdMetrics. The former makes
development a little easier in some circumstances. The latter we'll
use in production.

This adds testing infrastructure to make it easy to verify specific
metrics are generated during execution of bits of code.