Bug 1342619: Make boto3 client creation retryable
This reworks the retry code so it's easier to work with. It also changes boto3
client creation so that it gets retried a few times if it fails.

The theory here is that in Antenna -stage, the boto3 lib does an HTTP request to
get the credentials, gets rate-limited and then fails and that hoses Antenna's
ability to start up. Assuming the theory is right, this will alleviate that a
bit.