Bug 1342606: Move health stats to a heartbeat
Previously, Antenna would emit health stats for the save queue size and active
worker pool size as statsd data when /__heartbeat__ was handled. However, that
only gets handled by one process on the node, so the data we're getting back
from a node isn't the whole story and we won't get back data from orphaned
processes that aren't handling HTTP requests.

This changes the architecture so that BreakpadResource has a heartbeat coroutine
that sleeps 30 seconds and then emits stats.