Redo crashmoving
This renames some things to "crashmover" and related names. This should make it
easier to understand in the context of the current Socorro collector that this
replaces.

This changes the number of concurrent crashmovers running. Previously, we had 10
running. But it was actually (10 * number-of-processes) which is 50 per node and
they were all fighting for upload bandwidth. This changes that to 1 so there are
at most 5 per node. That should work a *lot* better.