Adjust crashmover spawning
Previously, a crashmover coroutine would spawn after an on_post. However, for
hand-wavey reasons, maybe a coroutine could "die" and if there's not an on_post
right afterwards, another one won't get spawned and Antenna might sit around for
a bit and not save something.

Anyhow, this adjusts things so that the heartbeat also kicks off coroutines if
appropriate.