Modify heartbeat system to keep Antenna alive if crashes exist
This modifies the Antenna heartbeat infrastructure such that it'll try to stay
alive if there are still crashes in the queue to be saved.

The way it "works" is by having the heartbeat coroutine exist in Gunicorn's
incoming connections pool. That's super tricky to do since that pool isn't
really available anywhere, so we have to do some gross stuff to find it.

Anyhow, this is using things that aren't likely to change soon, so I think we're
ok for now. Plus as far as I can tell, there's no other way to do this.