Bug 1347185: fix s3 connection problems
This changes the code to retry ValueError when trying to build a client. We get
a ValueError when boto3 uses the wrong endpoint.

Our theory is that it tries to get credentials and fails for some reason, then
degrades to this endpoint and fails.

Now if it hits this scenario (which seems to happen a lot when we start a new
node), it'll retry 5 times over 10 seconds (might not be enough--we'll find
out).

This also logs a line if it sent something to Sentry.