Fix the issue where nginx just drops the connection
If nginx exceeds the timeout when waiting for the rest of the request, it'll
just drop the connection. If the systemtest is connecting to the ELB, the ELB
turns that into an HTTP 504. If the systemtest is connecting directly to nginx,
then nginx just drops the connection and the systemtest fails.

This fixes the systemtest for the nginx case.