fix bug 1476941: reject unsupported products
This switches the throttler to reject unsupported products. We'll add
products to this list over time as requested.

This also cleans up the throttler. It gets rid of the weirdness around
conditions, so now all conditions are functions. It gets rid of the
one regexp--we could do the same thing without a regexp. It adds
documentation.