bug 1476941: comment out uuid throttle-override
This comments out the code that checks the accept/defer character of the
uuid in the crash report for crash reports that are coming in with a
uuid.

In the case where the crash report already has a uuid (or legacy
processing instructions), then it's already been accepted/defered
by prod and is being sent to stage to help us test out stage.

We're going to comment this out for now so we can test throttling changes.