fix bug 1480936: switch DEFER to REJECT
This isn't quite "switch to REJECT by default". What it does is adjust
the rule set such that rejection rules are all at the beginning of
the rule set and then throttling rules (which now accept or reject),
and then everything that's made it through that should get accepted.

There's one caveat to that. I kept the "is this an alpha, beta, esr"
and the "is this nightly" rules because throttle results get sent
to metrics aggregation and I happen to use that graph to see whether
a sudden influx of crashes is related to Firefox beta. Might make
sense to add a separate metrics pass that gathers information
about incoming crash reports from the collector. Then we could remove
those rules.