bug 1531865: redo scaffolding to match Socorro
This redoes Antenna configuration and scaffolding to match how Socorro
works.

This eliminates the need for prefacing things with ANTENNA_ENV=xyx.env which
simplifies command lines and examples.

This moves the dev.env thing into a docker/config/local_dev.env since that's
what people are likely to be using in a local dev environment.

This adds a my.env override file for handling overriding variables in the
Makefile as well as in the app.