bug 1527343: redo crashmover as a state machine; remove dead code
This removes all the load_crash_data code because it's not used anywhere.

This redoes the crashmover using a single queue with crash reports
having multiple states. This also redoes the crashmover_save() flow
to match how publishing works and moves the responsibility of saving
things in the right order to S3 to the S3 crashstorage.