bug 1359147: fix antenna startup bucket check
This fixes Antenna startup so that it verifies it can write to the S3 bucket.
It keeps the /__healthcheck doing a HEAD on the bucket. The figuring being
that if Antenna has started up and could write at that point, it's possible
but unlikely enough that that will change that it's not worth testing
explicitly.