bug 1618201: fix gunicorn worker class to be configurable
This was using the old gunicorn worker class that we created to handle
making grpcio aware we were using gevent. Now that we're not using
Pub/Sub, we aren't using grpcio and thus we don't need that anymore and
can go back to the gevent worker.

The infrastructure sets the GUNICORN_WORKER_CLASS variable, so we should
use that.