bug 1619644: handle None values in throttler
It's now possible for the collector to get None values for crash
annotations. None is not a string which makes some of the throttler
rules sad.

This adds a "safe_get" function which gets the value from the crash and
forces it to be a string. Then we can use string functions on it.