bug 1687987: remove legacy_processing from raw crash
This removes the legacy_processing item that the collector adds to the
raw crash. We used to have three possible throttle results: ACCEPT,
DEFER, and REJECT and needed to later distinguish between ACCEPT and
DEFER.

We no longer return DEFER as a throttle result and the legacy_processing
field isn't useful.