bug 1687987: remove timestamp from raw_crash
We had a timestamp as a seconds since epoch float in the raw crash. We
were using it to calculate how long it took to receive, parse, and save
the raw crash. We can figure that metric out from the
submitted_timestamp which is also added, so this removes timestamp.