bug 1756223: fix lint and test CI jobs
This undoes the "touch .docker-build" line which wasn't working between
runs.

This tweaks how we run lint and verify_reqs. They should be running in
the right container with no volume mounting (doesn't work in CI) and no
deps. Also, the lint job makes sure it has a my.env even though it's not
used.

This increases the urlwait in the test code--localstack s3 sometimes
takes too long to come up.

This adds yaml anchors to reduce some of the repeated bits in the
config.yml file.