bug 1562641: fix multipart handling
This removes the part where the loop ignores parts with the name
"dump_checksums". That's something the collector adds, but it's only
pertinent for the stage collector that's accepting re-packaged crash
reports from prod. Instead of handling that in the collector code, I
moved handling that to the socorro-submitter.

This also fixes regular parts and removes the sanitizing code. That
dropped - which broke the Add-ons annotation.