bug 1757815: fix systemtest SSLError issue
Previously, the systemtests would handle the case where the server
severs the connection, but requests wasn't able to see the HTTP response
because it was busy writing by catching that ConnectionError exception
and treating it as if the server returned an HTTP 413.

When we upgraded from Python 3.9 to 3.10, the underlying code changed
somehow so instead of getting a reqeuests.exception.ConnectionError, the
code now gets a requests.exception.SSLError. It's the same situation,
but the code reacts differently to it.

This updates the sleight-of-hand and the related comment.