bug 1764571: add sentry scrubbing
This overhauls our sentry support by switching to Fillmore to provide
scrubbing infrastructure and better safe set_up_sentry code.

This adds an integration test so if anything changes in sentry-sdk or
related things, we'll know and can look at the Sentry event again for
additional things we need to scrub.

Because this uses the SentryWsgiMiddleware, I had to adjust a bunch of
test code and gunicorn hook code to traverse one more layer of wsgi app
to find the things we have in AntennaApp.

This also breaks out version/revision/release functions into a new
libdockerflow shim.