bug 1687987: switch to raw crash version 2
This switches to the version 2 of the raw crash structure wherein we
move most of the collector-added fields into a "metadata" section and
additionally add a "version" field denoting the raw crash structure
version.

This also reworks BreakpadSubmitterResource so it's passing around a
CrashReport and then generate the raw crash structure from that. That
makes it a little easier to see the whole structure since it's in one
place.