bug-1819628: reject incoming crash reports by Android_PackageName
This adds a map of product -> supported Android_PackageName values and a
throttler rule to reject incoming crash reports where the configuration
denotes a list of supported packagenames and either the
Android_PackageName is missing or it's not in the supported list.

The initial set of supported packagenames is provided in bug-1819628.