bug-1891047: clean up test_breakpad_resource.py
This fixes the `FakeCrashMover` so that it has the right method and if
that's called, it raises an error now.

This breaks up the monolithic `TestBreakpadSubmitterResourceIntegration`
into logical groups of tests:

* tests for extract payload
* test cleanup_crash_report
* test throttle_result
* integration tests

This also renames `test_submit_crash_report_reply` to
`test_submit_crash_report` and asserts that the necessary parts made it
through `BreakpadResource` through `CrashMover` to crash storage.