bug-1891047: fix test setup/teardown for support services
This changes test configuration so it uses different queues and buckets
than the local dev environment.

This fixes setup/teardown for s3 and gcs tests by adding x_client and
x_helper fixtures. Previously the gcs tests did this for every
test--it's better to pull it into a fixture.

This fixes rebuild_app to also reset verify_funs. Previously it didn't
do that, so storage and publish classes would stack up and then tests
would fail.

This fixes test data in two cases where we were passing in the wrong
things.