bug-1891047: add tests for noop classes, clean up load_crash
This adds tests for the NoOp classes.

This cleans up the `load_crash` code in crashstorage classes so it's
consistent around what it does when files are missing that shouldn't be
missing. Now it throws exceptions for these
shouldn't-ever-happen-in-test situations.

This updates documentation that talked about crashmover
coroutines--that's not a thing anymore.

This removes the `GUNICORN_WORKER_CONNECTIONS` documentation and usage.
We now use the sync worker, so we don't need this anymore.

This also deletes `tests/unittest/test_crashstorage.py` which doesn't
test anything that `test_breakpad_resource.py` doesn't already tests.