bug-1898345: fix metrics key prefix
In bec86e31, I removed setting the statsd_prefix because I thought we
were adding "socorro.collector" in telegraf--not in DatadogMetrics
markus backend. That was wrong and I broke stats in Socorro stage AWS.

This fixes that by setting "socorro" in the MetricsInterface per our
convention and then fixing the keys to include "collector".

Then I discovered one of the tests was busted, so I fixed that. I also
added metrics assertions to test_submit_crash_report.

The host value can change, so I wrote an AnyTagValue class that'll let
us assert a tag was emitted with whatever value it has.